requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
Werkzeug==3.0.1
gunicorn==21.2.0